Ajustes de interface
Ajustando alguns elementos da interface e distribuindo melhor os demais recursos em arquivos diferentes.

+++ main.py
@@ -2,15 +2,39 @@
 # FILE OF MAIN INTERACTIONS #
 #############################
 
-from controllers.gameProjectController import *
+from src.controllers.defineProject import *
+from src.controllers.getProject import *
+from src.controllers.alterProject import *
+
 from desktop.interface import *
 
 from time import sleep
 
+gameDescriptions = [
+    'Resumo',
+    'Pontos Fortes',
+    'Pontos Fracos',
+    'Oportunidades',
+    'Ameaças',
+    'Objetivo/impacto',
+    'Baseado em história',
+    'Assetes Fundamentais',
+    'Animações/cinemáticas',
+    'Níveis Fundamentais',
+    'Funcionalidades de Rede',
+    'Sonoplastia',
+    'Jogabilidade Principal',
+    'Mecânicas Secundárias',
+    'Interfaces',
+    'Paleta1',
+    'Paleta2',
+    'Tempo Médio de Sessão'
+]
+
 while True:
     choice = menu(['Listar Documentos', 'Criar Novo Documento', 'Atualizar Dados de Documento', 'Deletar Documento de Jogo', 'Sair'])
     if choice == 1:
-        getGameProject()
+        getGameProjects()
         
     elif choice == 2:
         header('NOVO PROJETO DE GAME')
@@ -21,7 +45,7 @@ while True:
             'strongs': str(input('Pontos Fortes: ')),
             'weaks': str(input('Pontos Fracos: ')),
             'oportunities': str(input('Oportunidades: ')),
-            'theads': str(input('Ameaças: ')),
+            'threads': str(input('Ameaças: ')),
             'objective': str(input('Objetivo/impacto: ')),
             'history': str(input('Baseado em história? ')),
             'assets': str(input('Assets Fundamentais: ')),
@@ -39,13 +63,33 @@ while True:
         createGameProject(project)
         
     elif choice == 3:
-        cod = int(input('Digite o Código do Documento que deseja Alterar: '))
-        description = str(input('Altere as informações do Resumo: '))
-        updateGameProject(cod, description)
-        
+        if confirm(): 
+            cod = readInt('Digite o Código do Documento que deseja Alterar: ')
+               
+            gameInfos = getGameProject(cod)
+            
+            infosToUpdate = list()
+            
+            if gameInfos != None:
+                for i in range(1, len(gameInfos)):
+                    print(f'{i}. {gameDescriptions[i-1]}: {gameInfos[i]}')
+                    
+                    updatedInfo = str(input('Mudar para: '))
+                    infosToUpdate.append(updatedInfo)
+                    
+                updateGameProject(cod, infosToUpdate)
+            else:
+                print('Nenhum Documento com esse Código foi encontrado!')
+        else:          
+            print('Operação cancelada.')
+            
     elif choice == 4:
-        cod = int(input('Digite o Código do Documento que deseja Deletar: '))
-        deleteGameProject(cod)
+        if confirm():
+            cod = readInt('Digite o Código do Documento que deseja Deletar: ')
+            deleteGameProject(cod)
+        else:          
+            print('Operação cancelada.')
+        
     elif choice == 5:
         print('Saindo do sistema...Até logo!')
         break
